Move hooks to own folder and update imports
