+++ hooks/useDocumentData.js
@@ -1,4 +1,5 @@
 import useSWR from 'swr';
+
 import { fetcher } from '../utils/fetcher';
 
 export function useDocumentData(user, invoiceId) {
@@ -11,20 +12,6 @@ export function useDocumentData(user, invoiceId) {
   if (data) {
     data.invoice.id = invoiceId;
   }
-
-  return {
-    data,
-    loading: !error && !data,
-    error,
-  };
-}
-
-export function useCollectionData(user) {
-  const { data, error } = useSWR(
-    user ? ['/api/invoices', user.token] : null,
-    fetcher,
-    { refreshInterval: 5000 }
-  );
 
   return {
     data,
